Fix output file management (#31)
* output directory

* add file existence check

---------

+++ utils.py
@@ -27,9 +27,9 @@ class ProcessHandle:
         if self.stdout_file is not None:
             self.stdout_file.close()
 
-        if self.stdout_filename is not None:
+        if self.stdout_filename is not None and os.path.exists(self.stdout_filename):
             os.remove(self.stdout_filename)
-        if self.stderr_filename is not None:
+        if self.stderr_filename is not None and os.path.exists(self.stderr_filename):
             os.remove(self.stderr_filename)
 
         countdown = force_kill_after
